refactor job details crash
