+++ presentation/ui/src/main/java/com/cupcake/ui/base/ViewPagerAdapter.kt
@@ -1,11 +1,11 @@
-package com.cupcake.ui.posts
+package com.cupcake.ui.base
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-class ViewPagerPostsAdapter(
+class ViewPagerAdapter(
     fragmentManager: FragmentManager,
     fragmentList: List<Fragment>,
     lifecycle: Lifecycle
